fix autocomplete and combobox
