+++ src/components/GAutocomplete/GAutocompleteFactory.js
@@ -71,81 +71,6 @@ export function getInputEventHandlers(props, context, state, selections, selecte
   }
 
 }
-
-export function genTextFieldScopedSlot(props, context, selections, onChipCloseClick, isDirty, isValidInput, labelClasses, labelStyles, validateText, state, hintClasses, errorMessages, clearSelection) {
-  const textFieldScopedSlots = {
-    clearableSlot: ({iconColor}) =>
-        <GIcon vOn:click={clearSelection} vShow={isDirty.value && props.clearable}
-               color={iconColor}>{props.clearIcon}</GIcon>,
-    appendInner: ({iconColor}) =>
-        <GIcon color={iconColor}>arrow_drop_down</GIcon>,
-    inputSlot: ({inputErrStyles}) =>
-        <div class="g-tf-input" style={[{'color': '#1d1d1d'}, inputErrStyles]}>
-          {props.multiple ? genMultiSelectionsSlot() : genSingleSelectionSlot()}
-        </div>,
-    label: () => <label for="input" class={["g-tf-label", labelClasses.value]}
-                        style={labelStyles.value}>{props.label}</label>,
-    inputMessage: () => [<div v-show={props.counter} class={{
-      'g-tf-counter': true,
-      'g-tf-counter__error': !isValidInput.value
-    }}>{validateText.value.length}/{props.counter}</div>,
-      isValidInput.value ? <div class={["g-tf-hint", hintClasses.value]}>{props.hint}</div>
-          : <div class="g-tf-error">{errorMessages.value}</div>
-    ]
-  }
-  const genMultiSelectionsSlot = () => {
-    if (props.chips || props.smallChips || props.deletableChips || props.allowDuplicates) {
-      return selections.value.map((item, index) => <GChip small={props.smallChips}
-                                                          close={props.deletableChips}
-                                                          vOn:close={() => onChipCloseClick(index)}>{item}
-      </GChip>)
-    }
-
-    return selections.value.map(function (item, index) {
-          if (index === selections.value.length - 1) return <div
-              style={{'color': state.lastItemColor, 'padding-right': '5px'}}>{item}</div>
-          return <div style={{'padding-right': '5px'}}>{item + ', '} </div>
-        }
-    )
-  }
-  const genSingleSelectionSlot = () => {
-    if ((props.chips || props.smallChips || props.deletableChips) && selections.value) {
-      return <GChip small={props.smallChips} close={props.deletableChips}
-                    vOn:close={() => onChipCloseClick()}>{selections.value}</GChip>
-    }
-  }
-  return textFieldScopedSlots
-
-}
-
-export function genList(props, options, selectedItem, showOptions, context, selections, state) {
-  const onClickItem = () => {
-    setSearch(props, context, selections, state)
-    showOptions.value = props.multiple
-  }
-  return <GList
-      {...{
-        props: {
-          items: options.value,
-          'item-title': props.itemText,
-          mandatory: true,
-          allowDuplicates: props.allowDuplicates,
-          multiple: props.multiple,
-          selectable: true,
-          inMenu: true,
-          value: selectedItem.value,
-        },
-        on: {
-          'click:item': onClickItem,
-          input: e => selectedItem.value = e,
-        },
-      }
-      }
-      ref="list"
-  />
-
-}
-
 export function resetSelectionsDisplay(state) {
 
   state.pressDeleteTimes = 0
